// src/types/basicInfo.ts
export interface BasicInfo {
    fullName: string;
    DOB: string;
    Age: string;
    Gender: "Male" | "Female" | "Other";
    phoneNumber: string;
    email: string;
    HouseAddress: string;
    EmergencyNumber: string;
    NextOfKinName: string;
    NextOfKinGender: "Male" | "Female" | "Other";
    NextOfKinPhoneNumber: string;
    NextOfKinEmailAddress: string;
}